api/requirements.txt
fastapi
uvicorn
pymupdf
Pillow
numpy
opencv-python-headless